build.yml: Set CPPFLAGS, LDFLAGS, and XML_CATALOG_FILES possible for Homebrew locations.
On macOS, Intel versus Apple silicon has created a bit of a mess as
far as Homebrew is concerned. On the former, Homebrew formulae are
installed in /usr/local and on the latter, in /opt/homebrew. There's
no way to know apriori on which architecture the job will
run. Consequently, we attempt to handle both. In the future, perhaps
we can key off `uname -m`.